🏗️ Move router and placer into layout_synthsizer subdir
