+++ src/na/zoned/layout_synthesizer/placer/AStarPlacer.cpp
@@ -8,7 +8,7 @@
  * Licensed under the MIT License
  */
 
-#include "na/zoned/placer/AStarPlacer.hpp"
+#include "na/zoned/layout_synthesizer/placer/AStarPlacer.hpp"
 
 #include "ir/Definitions.hpp"
 #include "na/zoned/Architecture.hpp"
