+++ test/na/zoned/test_vertex_matching_placer.cpp
@@ -8,7 +8,7 @@
  * Licensed under the MIT License
  */
 
-#include "na/zoned/placer/VertexMatchingPlacer.hpp"
+#include "na/zoned/layout_synthesizer/placer/VertexMatchingPlacer.hpp"
 
 #include <cstddef>
 #include <gmock/gmock-matchers.h>
